Optional is Present

+++ src/main/java/dsi/senior/cnf/JwtTokenUtil.java
@@ -25,7 +25,7 @@ public class JwtTokenUtil implements Serializable {
 	 */
 	private static final long serialVersionUID = 1L;
 
-	public static final long JWT_TOKEN_VALIDITY = 5 * 60 * 60;
+	public static final long JWT_TOKEN_VALIDITY = (long) 5 * 60 * 60;
 
 	@Value("${jwt.secret}")
 	private String secret;
